+++ R/process.R
@@ -14,6 +14,7 @@ NULL
 #' ```
 #' p <- process$new(command = NULL, args,
 #'                  stdin = NULL, stdout = NULL, stderr = NULL,
+#'                  pty = FALSE, pty_options = list(),
 #'                  connections = list(), poll_connection = NULL,
 #'                  env = NULL, cleanup = TRUE, cleanup_tree = FALSE,
 #'                  wd = NULL, echo_cmd = FALSE, supervise = FALSE,
@@ -95,6 +96,21 @@ NULL
 #'     `"|"`: create a connection for it; `"2>&1"`: redirect it to the
 #'     same connection (i.e. pipe or file) as `stdout`. `"2>&1"` is a
 #'     way to keep standard output and error correctly interleaved.
+#' * `pty`: Whether to create a pseudo terminal (pty) for the background
+#'     process. This is currently only supported on Unix systems.
+#'     If it is `TRUE`, then the `stdin`, `stdout` and `stderr` arguments
+#'     must be `NULL`. If a pseudo terminal is created, then processx
+#'     will create pipes for standard input and standard output. There is
+#'     no separate pipe for standard error, because there is no way to
+#'     distinguish between stdout and stderr on a pty. Note that the
+#'     standard output connection of the pty is _blocking_, so we always
+#'     poll the standard output connection before reading from it using
+#'     the `$read_output()` method. Also, because `$read_output_lines()`
+#'     could still block if no complete line is available, this function
+#'     always fails if the process has a pty. Use `$read_output()` to
+#'     read from ptys.
+#' * `pty_options`: Unix pseudo terminal options, a named list. see
+#'     [default_pty_options()] for details and defaults.
 #' * `connections`: A list of connections to pass to the child process.
 #'     This is an experimental feature currently.
 #' * `poll_connection`: Whether to create an extra connection to the process
@@ -417,16 +433,16 @@ process <- R6::R6Class(
   public = list(
 
     initialize = function(command = NULL, args = character(),
-      stdin = NULL, stdout = NULL, stderr = NULL, connections = list(),
-      poll_connection = NULL, env = NULL, cleanup = TRUE,
-      cleanup_tree = FALSE, wd = NULL, echo_cmd = FALSE, supervise = FALSE,
-      windows_verbatim_args = FALSE, windows_hide_window = FALSE,
-      encoding = "",  post_process = NULL)
+      stdin = NULL, stdout = NULL, stderr = NULL, pty = FALSE,
+      pty_options = list(), connections = list(), poll_connection = NULL,
+      env = NULL, cleanup = TRUE, cleanup_tree = FALSE, wd = NULL,
+      echo_cmd = FALSE, supervise = FALSE, windows_verbatim_args = FALSE,
+      windows_hide_window = FALSE, encoding = "",  post_process = NULL)
 
       process_initialize(self, private, command, args, stdin,
-                         stdout, stderr, connections, poll_connection,
-                         env, cleanup, cleanup_tree, wd, echo_cmd,
-                         supervise, windows_verbatim_args,
+                         stdout, stderr, pty, pty_options, connections,
+                         poll_connection, env, cleanup, cleanup_tree, wd,
+                         echo_cmd, supervise, windows_verbatim_args,
                          windows_hide_window, encoding, post_process),
 
     finalize = function() {
@@ -590,6 +606,8 @@ process <- R6::R6Class(
     stdin = NULL,         # stdin argument or stream
     stdout = NULL,        # stdout argument or stream
     stderr = NULL,        # stderr argument or stream
+    pty = NULL,           # whether we should create a PTY
+    pty_options = NULL,   # various PTY options
     pstdin = NULL,        # the original stdin argument
     pstdout = NULL,       # the original stdout argument
     pstderr = NULL,       # the original stderr argument
@@ -729,4 +747,18 @@ process_close_connections <- function(self, private) {
       .Call(c_processx_connection_close, p)
     }
   }
+}
+
+#' Default options for pseudo terminals (ptys)
+#'
+#' @return Named list of default values of pty options.
+#'
+#' Options and default values:
+#' * `echo` whether to keep the echo on the terminal. `FALSE` turns echo
+#'   off.
+#'
+#' @export
+
+default_pty_options <- function() {
+  list(echo = FALSE)
 }