+++ src/init.c
@@ -8,9 +8,11 @@ void R_init_processx_win();
 void R_init_processx_unix();
 SEXP processx__unload_cleanup();
 SEXP run_testthat_tests();
+SEXP processx__echo_on();
+SEXP processx__echo_off();
 
 static const R_CallMethodDef callMethods[]  = {
-  { "processx_exec",               (DL_FUNC) &processx_exec,              14 },
+  { "processx_exec",               (DL_FUNC) &processx_exec,              16 },
   { "processx_wait",               (DL_FUNC) &processx_wait,               2 },
   { "processx_is_alive",           (DL_FUNC) &processx_is_alive,           1 },
   { "processx_get_exit_status",    (DL_FUNC) &processx_get_exit_status,    1 },
@@ -50,6 +52,8 @@ static const R_CallMethodDef callMethods[]  = {
   { "processx_disable_crash_dialog",  (DL_FUNC) &processx_disable_crash_dialog,   0 },
   { "processx_base64_encode", (DL_FUNC) &processx_base64_encode, 1 },
   { "processx_base64_decode", (DL_FUNC) &processx_base64_decode, 1 },
+  { "processx__echo_on", (DL_FUNC) &processx__echo_on, 0 },
+  { "processx__echo_off", (DL_FUNC) &processx__echo_off, 0 },
 
   { NULL, NULL, 0 }
 };