+++ src/processx.h
@@ -38,8 +38,8 @@ extern "C" {
 /* API from R */
 
 SEXP processx_exec(SEXP command, SEXP args,
-		   SEXP std_in,  SEXP std_out, SEXP std_err,
-		   SEXP connections, SEXP env,
+		   SEXP std_in,  SEXP std_out, SEXP std_err, SEXP pty,
+		   SEXP pty_options, SEXP connections, SEXP env,
 		   SEXP windows_verbatim_args,
 		   SEXP windows_hide_window, SEXP private_, SEXP cleanup,
 		   SEXP wd, SEXP encoding, SEXP tree_id);
@@ -101,6 +101,7 @@ typedef struct {
   int windows_verbatim_args;
   int windows_hide;
   const char *wd;
+  int pty_echo;
 } processx_options_t;
 
 #ifdef __cplusplus